+++ va416xx-hal/CHANGELOG.md
@@ -6,14 +6,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased]
+# [unreleased]
 
-## [v0.2.0]
+# [v0.2.0]
 
 - Documentation improvements
+- Improved UART typing support: Validity of passed pins is now checked properly
+
+## Changed
+
+- Added `va41620`, `va41630`, `va41628` and `va41629` device features. A device now has to be
+  selected for HAL compilation to work properly
+
+## Fixed
+
 - Small fixes and improvements for ADC drivers
+
+## Added
+
 - Added basic DMA driver
 
-## [v0.1.0] 2024-07-01
+# [v0.1.0] 2024-07-01
 
 - Initial release with basic HAL drivers